rename UpdateEntry -> updateEntry
